Data Structures P6 - Maze Master
Implements multidimensional arrays
The idea for this project is similar to old text-based adventure games in which a player is presented with various options through various screens of text.
In the case of your project, you have been trapped in a maze and must figure out how to navigate a maze in order to escape.
The maze is composed of rooms. In order to escape, the player needs to travel through each room of the maze, gathering keys and objects along the way
that will aid in their escape.