Data Structures P7 - Racko Game
This project is meant to help you familiarize yourself with sequential data structures. You will be recreating the card game of 'Racko', which is a game that involves rearranging your hand of cards in order to have an increasing sequence. Each card just has a number. We will be making a four player version of the game, where some players can be computer controlled. A user's moves are decided by the user's input, while the computer's moves will be decided by your program (you determine the strategy!). You will be using a combination of linked lists and arrays for doing this assignment.