Data Structures P8 - Boggle Master
This project is meant to help you familiarize yourself with backtracking algorithms and the usage of hashing data structures. You will be recreating the game of 'Boggle', which is a game that involves identifying as many words as possible from a 4 x 4 grid of letter dice. Each dice has 6 letters on it. This game will be single player with the player seeking to identify all words in the grid.